Working value cells.
