+++ main.py
@@ -1,78 +1,105 @@
-from dataclasses import dataclass
+import string 
+from typing import Optional
 
+class FormulaCell:
 
-@dataclass
-class HistoricalData:
-    interest_rate: list[float]
-    wacc: float
-    net_sales: list[float]
-    cogs: list[float]
-    discount: list[float]
-    operating_expense: list[float]
-    interest_expense: list[float]
-    taxes: list[float]
-    cash: list[float]
-    ar: list[float]
-    inventory: list[float]
-    ppne: list[float]
-    np_bank: list[float]
-    np_holtz: list[float]
-    np_trade: list[float]
-    ap: list[float]
-    accrued_expenses: list[float]
-    term_loan_current_portion: list[float]
-    term_loan: list[float]
-    net_worth: list[float]
+    def __init__(self, formula: str) -> None:
+        self._formula = formula
+        self._value = None
 
+    def evaluate(self) -> None:
+        pass
 
-@dataclass
-class ForecastAssumptions:
-    sales_growth: list[float]
-    interest_rate: list[float]
-    payables_period: list[float]
-    net_sales: float
+    def get_value(self) -> None:
+        return self._value
 
+    def set_formula(self, formula: str) -> None:
+        self._formula = formula
+
+    def get_formulat(self) -> None:
+        return self._formula
+
+class ValueCell:
+
+    def __init__(self, value: Optional[float] = 0):
+        self._value = value
+
+    def set_value(self, value: float) -> None:
+        self._value = value
+
+    def get_value(self) -> None:
+        return self._value
 
 class SpreadSheet:
 
-    def __init__(self, data: HistoricalData, assumptions: ForecastAssumptions) -> None:
-        # Do a bunch of magic
-        plug: list[float] = None
-        return plug
+    def __init__(self, num_rows: int, num_cols: int) -> None:
+        self._cells = {
+            char: [ValueCell() for _ in range(num_rows)] for char in string.ascii_uppercase[0: num_cols + 1]
+        }
 
+    def get_cell(self, loc: str) -> ValueCell | FormulaCell:
+        col, row = loc[0], loc[1:]
+        return self._cells[col][row]
 
+    def set_cell(self, loc: str, cell: ValueCell | FormulaCell) -> None:
+        col, row = loc[0], loc[1:]
+        row = int(row) - 1
+
+        self._cells[col][row] = cell
+
+    def _evaluate(self) -> None:
+        graph = self._get_dependencies()
+        pass
+
+    def _get_dependencies(self) -> None:
+        pass
+
+    def to_string(self):
+        self._evaluate()
+        result = "   "
+        
+        # Add column headers
+        for col in sorted(self._cells.keys()):
+
+            result += f"{col:^5}"  # Center-align column headers with 5 spaces
+
+        result += "\n"
+        
+        # Add rows with row numbers and cell values
+        for row_idx in range(len(next(iter(self._cells.values())))):
+
+            result += f"{row_idx + 1:2d} "  # Add row number with 2 digits of space
+            
+            # Add cell values for each column
+            for col in sorted(self._cells.keys()):
+
+                cell = self._cells[col][row_idx]
+                value = cell.get_value()
+
+                if value is None:
+                    value = ""
+
+                result += f"{value:^5}"  # Center-align cell values with 5 spaces
+
+            result += "\n"
+        
+        return result
+            
 if __name__ == "__main__":
-    data = HistoricalData(
-        interest_rate=[11.0, 11.0, 0],
-        wacc=10.0,
-        net_sales=[2921.0, 3477.0, 4519.0],
-        cogs=[2202.0, 2634.0, 3, 424],
-        discount=[0.0, 0.0, 0.0],
-        operating_expense=[622.0, 717.0, 940.0],
-        interest_expense=[23.0, 42.0, 56.0],
-        taxes=[14.0, 16.0, 22.0],
-        cash=[43.0, 52.0, 56.0],
-        ar=[306.0, 411.0, 606.0],
-        inventory=[337.0, 432.0, 587.0],
-        ppne=[233.0, 262.0, 388.0],
-        np_bank=[0.0, 60.0, 390.0],
-        np_holtz=[0.0, 200.0, 100.0],
-        np_trade=[0.0, 0.0, 127.0],
-        ap=[213.0, 340.0, 376.0],
-        accrued_expenses=[42.0, 45.0, 75.0],
-        term_loan_current_portion=[20.0, 20.0, 20.0],
-        term_loan=[140.0, 120, 0, 100.0],
-        net_worth=[504.0, 372.0, 449.0],
+    sheet = SpreadSheet(5, 5)
+    print(sheet.to_string())
+
+    sheet.set_cell(
+        loc="A1",
+        cell=ValueCell(5)
     )
 
-    assumptions = ForecastAssumptions(
-        sales_growth=[None, 25.0, 25.0, 25.0],
-        interest_rate=[11.0, 11.0, 11.0, 11.0],
-        payables_period=[53.6, 53.6, 53.6, 53.6],
-        net_sales=5500.0,
+    print(sheet.to_string())
+
+    sheet.set_cell(
+        loc="B1",
+        cell=FormulaCell("A1 + 1")
     )
 
-    sheet = SpreadSheet(
-        data=data,
-        assumptions=assumptions
-    )
+    print(sheet.to_string())
+
